src/streamlit_app/requirements.txt
beautifulsoup4==4.12.3
qdrant-client==1.10.1
langchain-qdrant==0.1.1
langchain-openai==0.1.14
langchain-core==0.2.12
langchain==0.2.7
streamlit==1.36.0
python-dotenv==1.0.1
streamlit-google-auth==1.1.8

